requirements.txt
aniso8601==3.0.2
asn1crypto==0.24.0
astroid==2.1.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.5
dj-database-url==0.5.0
Django==2.1.5
django-cors-headers==2.4.0
django-dbbackup==3.2.0
django-filter==2.1.0
django-group-by==0.3.1
django-heroku==0.3.1
django-rest-knox==3.6.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-yasg==1.13.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
idna==2.7
inflection==0.3.1
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
Pillow==5.4.0
promise==2.2.1
psycopg2==2.7.6.1
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.2.2
pyOpenSSL==19.0.0
pytz==2018.7
PyYAML==3.13
requests==2.19.1
requirements==0.1
ruamel.yaml==0.15.87
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
swagger-spec-validator==2.4.3
uritemplate==3.0.0
urllib3==1.23
whitenoise==4.1.2
wincertstore==0.2
wrapt==1.10.11
